Add files via upload
These are forms for registration and changing information on accounts. I'm not sure if PHP is mobile compatible so I haven't taken the time to write any code for these. I'm still working on the account info but I could use some help writing JavaScript to show little explanation boxes next to some of the input fields, and something needs to be done so that one empty group for a secondary talent shows up every time you enter information for one (maybe add a limit to this, maybe about 5 or 10). Also, let me know if you like the table setup or not. If you have an idea of what to do instead let me know and give an example.